requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.31.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
